refactor(TimeRestController) : rest api 컨트롤러 분리
