+++ src/main/java/roomescape/controller/TimeController.java
@@ -1,48 +1,15 @@
 package roomescape.controller;
 
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.*;
-import roomescape.dto.TimeResponse;
-import roomescape.dto.TimeSaveRequest;
-import roomescape.service.ReservationTimeService;
-
-import java.net.URI;
-import java.util.List;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
+@RequestMapping("/time")
 public class TimeController {
 
-    private final ReservationTimeService reservationTimeService;
-
-    public TimeController(ReservationTimeService reservationTimeService) {
-        this.reservationTimeService = reservationTimeService;
-    }
-
-    @GetMapping("/time")
+    @GetMapping
     public String getAdminPage() {
         return "/admin/time";
-    }
-
-    @GetMapping("/times")
-    public ResponseEntity<List<TimeResponse>> getTimes() {
-        List<TimeResponse> responses = reservationTimeService.findAllTimes();
-
-        return ResponseEntity.ok(responses);
-    }
-
-    @PostMapping("/times")
-    public ResponseEntity<TimeResponse> createTime(@RequestBody TimeSaveRequest request) {
-        TimeResponse response = reservationTimeService.saveTime(request);
-
-        URI location = URI.create("/times/" + response.id());
-        return ResponseEntity.created(location).body(response);
-    }
-
-    @DeleteMapping("/times/{id}")
-    public ResponseEntity<Void> deleteTime(@PathVariable Long id) {
-        reservationTimeService.deleteTimeById(id);
-
-        return ResponseEntity.noContent().build();
     }
 }
